+++ FWApplication/Classes/FWApplication/Toolkit/UIKit/UITableView+FWApplication.h
@@ -13,6 +13,9 @@ NS_ASSUME_NONNULL_BEGIN
 // UITableView分类(Plain有悬停，Group无悬停)
 @interface FWTableViewWrapper (FWApplication)
 
+/// 是否启动高度估算布局，启用后需要子视图布局完整，无需实现heightForRow方法(iOS11默认启用，会先cellForRow再heightForRow)
+@property (nonatomic, assign) BOOL estimatedLayout UI_APPEARANCE_SELECTOR;
+
 /// 清空Grouped样式默认多余边距，注意CGFLOAT_MIN才会生效，0不会生效
 - (void)resetGroupedStyle;
 
@@ -83,32 +86,6 @@ NS_SWIFT_NAME(TableViewCellBackgroundView)
 
 // 延迟加载背景视图，处理section圆角、阴影等。会自动设置backgroundView
 @property (nonatomic, strong, readonly) FWTableViewCellBackgroundView *backgroundView;
-
-@end
-
-#pragma mark - FWTableViewWrapper+FWTemplateLayout
-
-/**
- 表格自动计算并缓存cell高度分类，布局必须完整，系统方案实现
- */
-@interface FWTableViewWrapper (FWTemplateLayout)
-
-/**
- 单独启用或禁用高度估算
- @note 启用高度估算，需要子视图布局完整，无需实现heightForRow方法；禁用高度估算(iOS11默认启用，会先cellForRow再heightForRow)
- 
- @param enabled 是否启用
- */
-- (void)setTemplateLayout:(BOOL)enabled UI_APPEARANCE_SELECTOR;
-
-// 缓存方式获取估算高度，estimatedHeightForRowAtIndexPath调用即可。解决reloadData闪烁跳动问题
-- (CGFloat)templateHeightAtIndexPath:(NSIndexPath *)indexPath;
-
-// 设置估算高度缓存，willDisplayCell调用即可，height为cell.frame.size.height。设置为0时清除缓存。解决reloadData闪烁跳动问题
-- (void)setTemplateHeight:(CGFloat)height atIndexPath:(NSIndexPath *)indexPath;
-
-// 清空估算高度缓存，cell高度动态变化时调用
-- (void)clearTemplateHeightCache;
 
 @end